Day 1 Part 1
**This Commit**
Is the not-super-Rust code golf-y way to do Part 1.

**Why?**
There was a trailing newline and I got the wrong answer so many times
that I got frustrated and kept rewriting it simpler and simpler. Once I
realized it was the newline I didn't feel like undoing the work to the
algorithm. Future work will optimize this.